Move absolute knobbing fully to mk1 driver
